1차발표자료 병합중

+++ FPI/src/main/java/com/example/fpi/domain/dto/user/UserDTO.java
@@ -12,7 +12,7 @@ public class UserDTO {
     private String phoneNumber;
     private String userImg;
     private String email;
-    private Long userCash;
+    private int userCash;
     private String userProApproval;
     private Long userStarRate;
     private String role;
